requirements.txt
Flask==2.1.1
scikit-learn==1.0.2
Pillow==10.2.0
numpy==1.26.4
joblib==1.1.0
